+++ src/lib.rs
@@ -22,7 +22,10 @@ mod presets;
 mod traits;
 
 pub use generic::{sieve_and_find, SieveIterator};
-pub use presets::{generate_prime_with_rng, generate_safe_prime_with_rng, is_prime_with_rng, is_safe_prime_with_rng};
+pub use presets::{
+    fips_is_prime_with_rng, fips_is_safe_prime_with_rng, generate_prime_with_rng, generate_safe_prime_with_rng,
+    is_prime_with_rng, is_safe_prime_with_rng,
+};
 pub use traits::{RandomPrimeWithRng, SieveFactory};
 
 #[cfg(feature = "default-rng")]