+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.7.0] - in development
+
+### Added
+
+- `LucasCheck::Regular` (the recommended one from the FIPS-186.5 standard). ([#72])
+- `hamzat::minimum_mr_iterations()` (to calculate the number of MR checks according to the FIPS-186.5 standard). ([#72])
+- Add `fips_is_prime_with_rng()` and `fips_is_safe_prime_with_rng()`. ([#72])
+
+
+[#72]: https://github.com/entropyxyz/crypto-primes/pull/72
+
+
 ## [0.7.0-pre.0] - 2025-02-22
 
 ### Changed